Fix slice.hpp when -DNDEBUG is *not* provided

+++ pythran/pythonic/types/slice.hpp
@@ -181,10 +181,10 @@ namespace pythonic
       assert(not(upper.is_none and lower.is_none));
       long len;
       if (upper.is_none) {
-        assert(nt2::bitofsign(step) != nt2::bitofsign(lower));
+        assert(nt2::bitofsign((long)step) != nt2::bitofsign((long)lower));
         len = -(long)lower;
       } else if (lower.is_none) {
-        assert(nt2::bitofsign(step) == nt2::bitofsign(upper));
+        assert(nt2::bitofsign((long)step) == nt2::bitofsign((long)upper));
         len = upper;
       } else
         len = upper - lower;
